backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pytest>=8.1
httpx>=0.27
python-jose[cryptography]>=3.3
bcrypt>=4.0
